fix(spec): widen index-manifest row count to i64
`IndexFileMeta._ROW_COUNT` is an Avro `long` in Java (`IndexFileMeta.SCHEMA`
field 3, `new DataField(3, "_ROW_COUNT", new BigIntType(false))`), but the Rust
field is `i32` and the index-manifest decoder narrows it:

    "_ROW_COUNT" => row_count = Some(read_long_field(cursor, field.nullable)? as i32),

An index file with more than `i32::MAX` rows therefore decodes to a silently
wrong, possibly negative, row count. The index manifest is written by Java and
read here, so nothing surfaces the truncation.

Carry the count as `i64` end to end: widen the field, drop the cast, and widen
the three index builders' `checked_row_count` helpers, which computed the count
into an `i32` and rejected anything larger before it could reach the manifest.
Range arithmetic there is now checked rather than relying on the narrowing
conversion to catch a bad range. The only remaining narrowing is at the native
Lumina builder boundary, which counts rows in an `i32`; the manifest keeps the
full width.

Everything else is mechanical adaptation of construction sites and literals
across the library, the fulltext-gated code, the tests and the C bindings.

+++ crates/paimon/src/table/pk_full_text_bucket_search.rs
@@ -369,7 +369,7 @@ mod tests {
             index_type: PK_FULL_TEXT_INDEX_TYPE.into(),
             file_name: file_name.into(),
             file_size: 1,
-            row_count: total as i32,
+            row_count: total,
             deletion_vectors_ranges: None,
             global_index_meta: Some(gim(7, frame(level, files))),
         }
